requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pymongo==4.10.1
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
email-validator==2.2.0
boto3==1.35.36
Pillow==10.4.0
reportlab==4.2.5